Extract shared tests
